hf_spoilage_model/requirements.txt
torch
torchvision
pillow
huggingface_hub==0.26.5
